requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-robots==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wincertstore==0.2
wrapt==1.11.2
